requirements.txt
flask
flask-cors
pillow
diffusers
transformers
torch
requests
python-dotenv
